backend/summaries/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
-e file:///home/rob/Code/english-vocabulary-tool/backend/common
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.9.0
    # via openai
openai==1.68.2
    # via -r requirements.in
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.1.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
